extract exam_page model and rename exam_page to exam_page_widget (#47)
