+++ lib/domain/models/exam/exam_page.dart
@@ -1,0 +1,13 @@
+import 'exam.dart';
+
+class ExamPage {
+  const ExamPage({
+    required this.exam,
+    required this.lVerticalText,
+    required this.rVerticalText,
+  });
+
+  final Exam exam;
+  final String? lVerticalText;
+  final String? rVerticalText;
+}
